Only check event conditions once per card.
Refactor the code so that we only check to see if the
event on a card is playable once per card.  This is necessary,
because some of the condition test require asking the user for
information. (What cards are in the discard pile for example)
And we do not want to ask more than once!