Keep track of countries where plots resolved
The Je Suis Charlie event is only playable if a plot
was resolved in a non-Muslim country in the previous plot
resolution phase.  This is not detected without haveing to
ask the user.

The US Bot no longer has to ask if it is the last card play of
the action phase.