Rework bot event triggering
If the user does Ops first, we must test the event triggering
condition of the bot associated card after the Ops because the
board state could change and affect whether or not the event
gets triggered.
Since we now have to check the event conditions multiple times,
we cache the result of any question that was asked of the user
when determining if the event is playable.  No need to ask
the same question twice!