Fix bugs in Trade Embargo and Iranian Elections
Trade Embargo was not allowing a reaction marker to be placed in a Shia-mix
country. This was due to incorrectly checking that ArabWinter was in effect.
Also, NTP Safeguards was not checking for the presence of the US Trade Embargo
marker as a block to its execution.

Iranian Elections was not decreasing funding by 1.