Add a second Boolean parameter to EventConditions
This determines if the event is as result of being triggered
when the human opponent has played the card.  This lets is
adjust conditions on whether or not the card event should occur.

The Boston Marathon event now uses the forTrigger parameter to
allow the event to occur even if the GWOT penalty is > 0 if the
event is due to being triggered.