Fold GameParams into GameState, update priorities
The GameParams class was only needed to work around Scala's
limit of 22 fields in case class constructors.  That limitation
was removed with the release of Scala 11 and since I am reworking
the save game file this is a good time to remove the the class
that is no longer needed.

Updated the priority lists to reflect the change in order concerning
Syria/Iran in Forever War vs Awakening.  We now change the priorities
based on the current game mode.