Fix bug in Civil War Attrition with Nigeria
Nigeria will convert to a Non Muslim country whenever it is
and Ally and contains no cells.

This can happen during Civil War Attrition.  The code for
Civil War Attrition now checks for this and stops trying to
process unresolved `hits` after the country has converted which
immediately ends the Civil War.