Fix corner case bug with #195 Taliban Resurgent
If the event targets Nigeria while it is an Ally and the event removes
the last cell cell from Nigeria, then Nigeria reverts to a NonMuslim country.
When the code subsequently tried to perform a Jihad in Nigeria (as per the event)
that cause the code to crash since Jihads cannot take place in NonMuslim countries.
We now check to make sure the target is still Muslim after removing the cells before
doing the Jihad.