Allow US Bot to play #163 Status Quo in Ally
The event instructions chart for this event reads:
Adversary then Neutral.  Priority to highest Resource.

We we disallowing the event in Ally countries.
But since most Regime Change countries are Ally countrys and
since the event ends the Regime change, it seems that it is in
the US Bots interest to play/trigger the event.