fix: remember Nigeria status when plot resolved
When a plot is resolved in a country we now
save the muslim/non-muslim status of the country
when the plot was resolved.  This is necessary
because when a plot/2 or plot/3 is resolved
in Nigeria, it changes from non-muslim to muslim.

If card 224 "Je Suis Charlie" is played in the
subsequent round it needs to know if a plot
was resolved in a non-muslim country.

Also updated the save file scheme to mimic
that used in Fitl and ColTwi.

Also improve UI a bit..