fix: handling of extra cells
The management of extra cells granted via
Training Camps/Al Baghdadi was not correct.

When removing the Al Bahdadi via another event
such as Bowling Green Massacre (353) we were
not updating the available extra cells.

Also both events are no longer in play, the cells
on the map should remain in play until removed,
but those in the available box should be removed
immediately.  This was not being done.