fix: bot travel from auto-recruit countries
Fix the bot code so that it does not attempt to
travel the last cell from an auto-recruit
country unless there are at least two other
auto-recruit countries with cells.
We we not taking into account cells that had
been used in previous travel attempts during
the same travel operation.