feat: jihadist bot enhancments
Travel:  The bot will now only prioritize
traveling a cell to a Fair country if there is
an adjacent cell that can travel there.

Plot: If there are no WMD plot available, the
bot will always select the highest number plot
available to place.  If there are WMD plots
available, then the selction is random.