fix: monitor drawn/discarded cards during turn
When Avenger is draw it must trigger
When an auto trigger event is drawn it must trigger

Also, we now alway remind the user to place the
Critical Middle card in the middle of the draw pile
rather than discard it.  This happens if it is played
for any reason or if it is discarded for any reason.