feat: Ask to continue game after Victory achieved
Instead of asking the user up front if they wish to
continue the game once Victory is achieved, we now
ask as the moment that Vicotory is first achieved.

We also no longer show the victory achieved message
continuously once the user has chosen to continue.
The vitory message is shown once only when victory is
first achieved.