fix: Awakening/reaction marker type choice
When carrying out Unassociated events that allow the player to
choose the type of marker to place we no longer assume the
US player always wants to place an Awakening marker and the
Jihadist player wants to place a Reaction marker.
There are some cases where the player may wish to do the
opposite.