feat: Allow for lapsing/1st plot markers
When a lapsing card is dran from the lapsing box
during a turn via event, the event is still in effect
until the end of the turn so we must allow for the
lapsing event to be marked even though the card is
no longer there.  Same goes for the 1st plot box.