fix: action order with triggered events
Now when the US player plays a 3 Op for their first card play
of an action phase, we prompt up front to ask if they wish
to do Reassessment.  Only then do we ask about event order.

Also, if an opponent associated event will not trigger
then we do not bother to ask at all.