fix: 94 The door of Itjihad was closed
This event targets countries that were improved to Fair or Good
in the current or previous action phase.

But it turns out that the country must also still be Fair or Good.
So we must filter out any that have been degraded to Poor/improved
since they were tested/improved.