feat: add enh bot difficulty levels
The difficulty levels determine when US associated events play by
the Enh Jihadist Bot are triggered:
Easy    The events always trigger and the printed Ops from the card
        are added to the Jihadist reserves
Medium  The events trigger when Jihadist reserves are at zero.
        The printed Ops from the card are added to Jihadist reserves.
        When Jihdist reserves > 0, the event does not trigger but
        the printed Ops are subtracted from Jihadist reserves.
Hard    The events are never triggered

Several US associated events will improved the governance of a
Muslim country.  If triggering one of these events could potentially
result in an automatic US victory, then this is treated as a special
case where the event is not triggered.