fix: Enh Jihad Major Jihad at Poor decision
When checking if Major Jihad is possible a Poor while
processing the EvO, we now check regarless of available Ops.
(ie. We check as if 3 Ops are available).
If this passes, then the next EvO box checks to see if there
are enough Ops on the current card.  If so we proceed with the
Major Jihad, otherwise we add the card's Ops to reserves.