fix: Allow US to add reserves after WoI test
The player can select an untested Muslim country for the target
of a War of Ideas operation.  It is possible that after testing
the player does not have enough Ops from the card to complete
the operation.  We now allow the user to expend their reserves
after seeing the result of this roll.